fix(core): update version events documentId
